+++ src/pages/UserProfile.jsx
@@ -3,9 +3,9 @@ import NavBar from "../components/navBar/NavBar";
 
 const UserProfile = () => {
   return (
-    <div>
+    <>
       <NavBar />
-    </div>
+    </>
   );
 };
