+++ src/styles/userProfileStyle/userInputStyle.js
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+export const InputWrapper = styled.div`
+  margin: 0.5rem 0;
+`;
+export const InfoLabel = styled.label`
+  margin-right: 0.5rem;
+`;
+export const InfoInput = styled.input`
+  width: 200px;
+  padding: 0.25rem 0.5rem;
+`;
